requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
httpx==0.27.0
lockfile==0.12.2